refactor: 도메인 구조 변경
도메인 클래스를 리팩터링하여 적절한 패키지로 이동하고 다른 엔터티와의 관계를 업데이트했습니다.
