--- src/main/java/com/twentyone/steachserver/domain/Student.java
@@ -1,23 +1,0 @@
-package com.twentyone.steachserver.domain;
-
-import jakarta.persistence.*;
-import java.time.LocalDateTime;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "students")
-@NoArgsConstructor
-public class Student {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @OneToOne
-    @JoinColumn(name = "login_credentials_id")
-    private LoginCredential loginCredential;
-
-    @Column(length = 30)
-    private String name;
-    private LocalDateTime createdAt;
-    private LocalDateTime updatedAt;
-}
