+++ src/main/java/com/twentyone/steachserver/domain/classroom/model/Classroom.java
@@ -1,24 +1,25 @@
-package com.twentyone.steachserver.domain.classroom;
+package com.twentyone.steachserver.domain.classroom.model;
 
 
-import com.twentyone.steachserver.domain.Lecture;
+import com.twentyone.steachserver.domain.lecture.model.Lecture;
 import jakarta.persistence.*;
 import lombok.*;
 
-
+@Getter(value = AccessLevel.PUBLIC)
+@Setter(value = AccessLevel.PRIVATE)
+@ToString
+@EqualsAndHashCode
 @Entity
 @Table(name = "classrooms")
 @NoArgsConstructor
 public class Classroom {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String sessionId;
 
     @OneToOne
     @JoinColumn(name = "lectures_id")
     private Lecture lecture;
-
-    // Getters and Setters
 }
