+++ src/main/java/com/twentyone/steachserver/domain/member/model/Admin.java
@@ -1,10 +1,15 @@
-package com.twentyone.steachserver.domain;
+package com.twentyone.steachserver.domain.member.model;
 
+import com.twentyone.steachserver.domain.member.auth.model.LoginCredential;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
-import lombok.NoArgsConstructor;
 
-@NoArgsConstructor
+import lombok.*;
+
+@Getter(value = AccessLevel.PUBLIC)
+@Setter(value = AccessLevel.PRIVATE)
+@ToString
+@EqualsAndHashCode
 @Entity
 @Table(name = "admins")
 public class Admin {
@@ -12,12 +17,15 @@ public class Admin {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
+    @Column(length = 30)
+    private String name;
+
     @OneToOne
     @JoinColumn(name = "login_credentials_id")
     private LoginCredential loginCredential;
 
-    @Column(length = 30)
-    private String name;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
+
+
 }
