+++ src/main/java/com/twentyone/steachserver/domain/member/model/Teacher.java
@@ -1,9 +1,15 @@
-package com.twentyone.steachserver.domain;
+package com.twentyone.steachserver.domain.member.model;
 
+import com.twentyone.steachserver.domain.member.auth.model.LoginCredential;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
-import lombok.NoArgsConstructor;
 
+import lombok.*;
+
+@Getter(value = AccessLevel.PUBLIC)
+@Setter(value = AccessLevel.PRIVATE)
+@ToString
+@EqualsAndHashCode
 @Entity
 @Table(name = "teachers")
 @NoArgsConstructor
@@ -12,14 +18,15 @@ public class Teacher {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
+    @Column(length = 30)
+    private String name;
+    private Integer volunteerTime;
+    private String pathQualification;
+
     @OneToOne
     @JoinColumn(name = "login_credentials_id")
     private LoginCredential loginCredential;
 
-    @Column(length = 30)
-    private String name;
-    private Integer volunteerTime;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private String pathQualification;
 }
