--- src/main/java/com/twentyone/steachserver/domain/quiz/QuizChoice.java
@@ -1,22 +1,0 @@
-package com.twentyone.steachserver.domain.quiz;
-
-import com.twentyone.steachserver.domain.quiz.model.Quiz;
-import jakarta.persistence.*;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "quiz_choices")
-@NoArgsConstructor
-public class QuizChoice {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @ManyToOne
-    @JoinColumn(name = "quizzes_id")
-    private Quiz quiz;
-
-    private Boolean isAnswer;
-    private String choiceSentence;
-}
-
