+++ src/main/java/com/twentyone/steachserver/domain/studentCurricula/model/StudentsCurricula.java
@@ -1,0 +1,45 @@
+package com.twentyone.steachserver.domain.studentCurricula.model;
+
+import com.twentyone.steachserver.domain.curricula.model.Curricula;
+import com.twentyone.steachserver.domain.member.model.Student;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter(value = AccessLevel.PUBLIC)
+@Setter(value = AccessLevel.PRIVATE)
+@ToString
+@EqualsAndHashCode
+@Entity
+@Table(
+        name = "students_curricula",
+        uniqueConstraints = {
+                @UniqueConstraint(
+                        name = "uk_student_curricula",
+                        columnNames = {"curricula_id", "students_id"}
+                )
+        }
+)
+public class StudentsCurricula {
+    @EmbeddedId
+    private StudentsCurriculaId id;
+
+    @ManyToOne
+    @MapsId("curriculaId")
+    @JoinColumn(name = "curricula_id")
+    private Curricula curricula;
+
+
+    @ManyToOne
+    @MapsId("studentId")
+    @JoinColumn(name = "students_id")
+    private Student student;
+
+    protected StudentsCurricula() {}
+
+    private StudentsCurricula(Student student, Curricula curricula) {
+        this.id = StudentsCurriculaId.createStudentsCurriculaId(student.getId(), curricula.getId());
+        this.student = student;
+        this.curricula = curricula;
+    }
+
+}
